etl/requirements.txt
clickhouse-driver==0.2.3
backoff==2.0.1
pydantic==1.9.1
python-dotenv==0.20.0
aiokafka==0.7.2